pyproject.toml
[tool.poetry]
name = "biocypher-project-template"
version = "0.1.0"
description = "Template for creating a BioCypher-driven knowledge graph"
authors = ["Sebastian Lobentanzer <sebastian.lobentanzer@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
biocypher = { git = "https://github.com/saezlab/BioCypher.git", branch = "main" }
polars = "^0.16.1"
pypath-omnipath = "^0.14.33"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
